+++ app/serializers/components/heading1.rb
@@ -1,0 +1,16 @@
+module Serializers
+  class Heading1 < Base
+
+    def initialize(person)
+      @person = person
+    end
+
+    def content
+      {
+        name:  "heading1",
+        data: "#{@person.full_name}"
+      }
+    end
+
+  end
+end
