+++ app/serializers/serializers.rb
@@ -3,6 +3,8 @@ require_relative './page'
 require_relative './pages/person_show_page'
 require_relative './pages/list_page'
 require_relative './components/contact'
+require_relative './components/heading1'
+require_relative './components/subheading'
 require_relative './components/image'
 require_relative './components/person'
 require_relative './components/related_links'
